+++ Railgun/Commands/Music/MusicPlaylist.cs
@@ -74,35 +74,6 @@ namespace Railgun.Commands.Music
                 await (Context.Channel as ITextChannel).SendStringAsFileAsync("Playlist.txt", output.ToString(), $"{Context.Guild.Name} Music Playlist ({playlist.Songs.Count} songs)");
                 await response.DeleteAsync();
             }
-
-            [Command("export"), BotPerms(ChannelPermission.AttachFiles), UserPerms(GuildPermission.ManageGuild)]
-            public async Task ExportAsync() {
-                var profile = Context.Database.ServerProfiles.GetOrCreateData(Context.Guild.Id);
-            	var data = profile.Music;
-
-                if (data == null || data.PlaylistId == ObjectId.Empty) {
-                    await ReplyAsync("There's no playlist data to export.");
-                    return;
-                }
-
-                var response = await ReplyAsync("Building Playlist data file. Standby...");
-                var output = new StringBuilder()
-                    .AppendFormat("# {0}'s Playlist.", Context.Guild.Name).AppendLine()
-                    .AppendFormat("# Generated At : {0}", DateTime.Now).AppendLine()
-                    .AppendLine("#")
-                    .AppendLine("# !!! DO NOT CHANGE/MODIFY THIS FILE !!! ")
-                    .AppendLine();
-                
-                var playlist = await SystemUtilities.GetPlaylistAsync(_musicService, data);
-
-                foreach (var song in playlist.Songs) {
-                    if (song.ProcessorId != "YOUTUBE") continue;
-                    output.AppendLine(song.ToString());
-                }
-
-                await (Context.Channel as ITextChannel).SendStringAsFileAsync($"playlist-data{SystemUtilities.FileExtension}", output.ToString());
-                await response.DeleteAsync();
-            }
         }
     }
 }
